Workaround applet crash due to bad config
See https://github.com/Kilian/f.lux-indicator-applet/issues/66#issuecomment-186770071 for details. I notice that the version of the code I'm editing here has a different list of color temps than the version I have installed locally, so I'm using a default value of 0 since that should always be valid. I'm sad to see that the 2300 temperature option is gone in the current version of the code :(